move getMarkdownImageSize to @utils/markdown
