+++ app/utils/markdown/index.ts
@@ -5,6 +5,8 @@ import {Platform, StyleSheet} from 'react-native';
 
 import {changeOpacity, makeStyleSheetFromTheme} from '@utils/theme';
 
+import {getViewPortWidth} from '../images';
+
 export function getCodeFont() {
     return Platform.OS === 'ios' ? 'Menlo' : 'monospace';
 }
@@ -211,3 +213,46 @@ export function switchKeyboardForCodeBlocks(value: string, cursorPosition: numbe
 
     return 'default';
 }
+
+export const getMarkdownImageSize = (
+    isReplyPost: boolean,
+    isTablet: boolean,
+    sourceSize?: SourceSize,
+    knownSize?: PostImage,
+) => {
+    let ratioW;
+    let ratioH;
+
+    if (sourceSize?.width && sourceSize?.height) {
+        // if the source image is set with HxW
+        return {width: sourceSize.width, height: sourceSize.height};
+    } else if (knownSize?.width && knownSize.height) {
+        // If the metadata size is set calculate the ratio
+        ratioW = knownSize.width > 0 ? knownSize.height / knownSize.width : 1;
+        ratioH = knownSize.height > 0 ? knownSize.width / knownSize.height : 1;
+    }
+
+    if (sourceSize?.width && !sourceSize.height && ratioW) {
+        // If source Width is set calculate the height using the ratio
+        return {width: sourceSize.width, height: sourceSize.width * ratioW};
+    } else if (sourceSize?.height && !sourceSize.width && ratioH) {
+        // If source Height is set calculate the width using the ratio
+        return {width: sourceSize.height * ratioH, height: sourceSize.height};
+    }
+
+    if (sourceSize?.width || sourceSize?.height) {
+        // if at least one size is set and we do not have metadata (svg's)
+        const width = sourceSize.width;
+        const height = sourceSize.height;
+        return {width: width || height, height: height || width};
+    }
+
+    if (knownSize?.width && knownSize.height) {
+        // When metadata values are set
+        return {width: knownSize.width, height: knownSize.height};
+    }
+
+    // When no metadata and source size is not specified (full size svg's)
+    const width = getViewPortWidth(isReplyPost, isTablet);
+    return {width, height: width};
+};
