+++ types/utils/markdown.d.ts
@@ -1,0 +1,7 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+type SourceSize = {
+    height?: number;
+    width?: number;
+}
